@@ -9,35 +9,35 @@
     "@jest/console": "^29.4.1",
     "@jest/reporters": "^29.4.1",
     "@jest/test-result": "^29.4.1",
-    "@trunkio/launcher": "^1.2.4",
+    "@trunkio/launcher": "^1.3.1",
     "@types/caller": "^1.0.2",
     "@types/debug": "^4.1.12",
-    "@types/jest": "^29.5.11",
+    "@types/jest": "^29.5.12",
     "@types/jest-specific-snapshot": "^0.5.9",
-    "@types/node": "^20.10.6",
-    "@typescript-eslint/eslint-plugin": "^6.18.0",
-    "@typescript-eslint/parser": "^6.18.0",
+    "@types/node": "^20.12.7",
+    "@typescript-eslint/eslint-plugin": "^7.7.0",
+    "@typescript-eslint/parser": "^7.7.0",
     "caller": "^1.1.0",
     "debug": "^4.3.4",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.5.2",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^27.6.1",
+    "eslint-plugin-jest": "^28.2.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prefer-arrow-functions": "^3.1.4",
-    "eslint-plugin-prettier": "^5.1.2",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-prefer-arrow-functions": "^3.3.2",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-simple-import-sort": "^12.1.0",
     "fast-sort": "^3.2.0",
     "jest": "^29.3.1",
     "jest-specific-snapshot": "^8.0.0",
-    "semver": "^7.3.8",
-    "simple-git": "^3.22.0",
-    "ts-jest": "^29.0.3",
+    "semver": "^7.6.0",
+    "simple-git": "^3.24.0",
+    "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.1.2",
-    "typescript": "^5.3.3",
-    "yaml": "^2.2.0"
+    "typescript": "^5.4.5",
+    "yaml": "^2.4.1"
   },
   "bundleDependencies": [
     "tests"
