@@ -3,7 +3,7 @@ import { osTimeoutMultiplier, skipOS, TEST_DATA } from "tests/utils";
 
 // Note that the first-time ruby/rufo download can sometimes take a while.
 // Ruby build is quite slow on Mac, so only run tests on linux for now
-jest.setTimeout(600000 * osTimeoutMultiplier); // 300s or 900s
+jest.setTimeout(600000 * osTimeoutMultiplier);
 
 // Ruby build is quite slow on Mac, so only run tests on linux for now
 customLinterCheckTest({
