@@ -1,7 +1,7 @@
 import { linterCheckTest } from "tests";
 import { osTimeoutMultiplier } from "tests/utils";
 
-jest.setTimeout(900000 * osTimeoutMultiplier); // 300s or 900s
+jest.setTimeout(900000 * osTimeoutMultiplier);
 
 // TODO(Tyler): We will eventually need to add a couple more test cases involving failure modes.
 linterCheckTest({ linterName: "kube-linter" });
