@@ -1,5 +1,15 @@
-import { linterCheckTest } from "tests";
+import semver from "semver";
+import { customLinterCheckTest } from "tests";
+import { TEST_DATA } from "tests/utils";
 
-linterCheckTest({
+const versionGreaterThanOrEqual = (a: string, b: string) => {
+  const normalizedA = a.replace("pmd_releases/", "");
+  const normalizedB = b.replace("pmd_releases/", "");
+  return semver.gte(normalizedA, normalizedB);
+};
+
+customLinterCheckTest({
   linterName: "pmd",
+  args: TEST_DATA,
+  versionGreaterThanOrEqual,
 });
