@@ -1,7 +1,7 @@
 import { customLinterCheckTest } from "tests";
 import { osTimeoutMultiplier, TEST_DATA } from "tests/utils";
 
-jest.setTimeout(600000 * osTimeoutMultiplier); // 300s or 900s
+jest.setTimeout(600000 * osTimeoutMultiplier);
 
 customLinterCheckTest({
   linterName: "renovate",
