@@ -1,7 +1,7 @@
 import { makeToolTestConfig, toolTest } from "tests";
 import { osTimeoutMultiplier, skipOS } from "tests/utils";
 
-jest.setTimeout(600000 * osTimeoutMultiplier); // 300s or 900s
+jest.setTimeout(600000 * osTimeoutMultiplier);
 
 toolTest({
   toolName: "renovate",
